+++ database/migrations/2022_01_23_205444_create_dropped_products_table.php
@@ -18,7 +18,7 @@ class CreateDroppedProductsTable extends Migration
             $table->integer('user_id');
             $table->integer('product_id');
             $table->integer('amount');
-            $table->float('price', 8, 2); // store the current price of the product
+            $table->integer('price'); // store the current price of the product
             $table->timestamps();
         });
     }
